pagina do conselheiro e integração com promisetracker

+++ web/lib/WebSMM/Controller/User.pm
@@ -8,9 +8,11 @@ BEGIN { extends 'Catalyst::Controller' }
 sub base : Chained('/root') : PathPart('user') : CaptureArgs(0) {
     my ( $self, $c ) = @_;
 
-    if ( !$c->user || !grep { /^user$/ } $c->user->roles ) {
+	if ( !$c->user && !grep { /^user$/ } $c->user->roles && !grep { /^counsil$/ } $c->user->roles && !grep { /^counsil_master$/ } $c->user->roles ) {
         $c->detach( '/form/redirect_error', [] );
     }
+
+
 
     my $api  = $c->model('API');
     my $form = $c->model('Form');
